Move datasetFromExperiment out of benchmark provider
